diskspace statistics are now a separate thread, to isolate the delays introduced by statvfs()
