+++ src/main.c
@@ -45,6 +45,7 @@ struct netdata_static_thread static_threads[] = {
     {"macos",              "plugins",   "macos",      1, NULL, NULL, macos_main},
 #else
     {"proc",               "plugins",   "proc",       1, NULL, NULL, proc_main},
+    {"diskspace",          "plugins",   "diskspace",  1, NULL, NULL, proc_diskspace_main},
 #endif /* __FreeBSD__, __APPLE__*/
     {"cgroups",            "plugins",   "cgroups",    1, NULL, NULL, cgroups_main},
     {"check",              "plugins",   "checks",     0, NULL, NULL, checks_main},
